ViralLoopsButton component
