+++ app/packs/src/components/top_bar/LoggedOutTopBar.jsx
@@ -1,27 +1,15 @@
-import React, { useEffect } from "react";
+import React from "react";
 import { useWindowDimensionsHook } from "src/utils/window";
 import ThemeContainer, { useTheme } from "src/contexts/ThemeContext";
 
 import { LogoLight, LogoDark, Logo } from "src/components/icons";
 import Button from "src/components/design_system/button";
+import ViralLoopsButton from "src/components/design_system/button/ViralLoopsButton";
 import { Sun, Moon } from "src/components/icons";
 
 export const LoggedOutTopBar = ({}) => {
   const { mobile } = useWindowDimensionsHook();
   const { simpleToggleTheme, mode } = useTheme();
-
-  useEffect(() => {
-    const script = document.createElement("script");
-
-    script.innerHTML = `!function(a,b,c,d,t){var e,f=a.getElementsByTagName("head")[0];if(!a.getElementById(c)){if(e=a.createElement(b),e.id=c,e.setAttribute("data-vrlps-ucid",d),e.setAttribute("data-vrlps-version","2"), e.setAttribute("data-vrlps-template", t),e.src="https://app.viral-loops.com/popup_assets/js/vl_load_v2.min.js",window.ub){jQuery=null,$=null;var g=a.createElement(b);g.src="https://code.jquery.com/jquery-2.2.4.min.js",f.appendChild(g)}f.appendChild(e);var h=a.createElement("link");h.rel="stylesheet",h.type="text/css",h.href="https://app.viral-loops.com/static/vl-loader.css",f.appendChild(h);var i=a.createElement("div");i.id="vl-overlay",i.style.display="none";var j=a.createElement("div");j.id="vl-loader",i.appendChild(j),a.addEventListener("DOMContentLoaded",function(b){a.body.appendChild(i);for(var c=a.getElementsByClassName("vrlps-trigger"),d=0;d<c.length;d++)c[d].removeAttribute("href"),c[d].onclick=function(){a.getElementById("vl-overlay").style.display="block"};var e=a.querySelectorAll("[data-vl-widget='popupTrigger']");[].forEach.call(e,function(b){var c=a.createElement("div");c.className="vl-embedded-cta-loader",b.appendChild(c)})})}}(document,"script","vrlps-js","wzibQEVPjtucE2keZEzIPOrVPcs","waitlist")`;
-    script.id = "ViralLoops";
-
-    document.body.appendChild(script);
-
-    return () => {
-      document.body.removeChild(script);
-    };
-  }, []);
 
   return (
     <div className="navbar-container">
@@ -40,12 +28,7 @@ export const LoggedOutTopBar = ({}) => {
           )}
         </div>
         <div className="d-flex" style={{ height: 34 }}>
-          <Button
-            type="primary-default"
-            className="vrlps-trigger"
-            onClick={() => null}
-            text="Join Waitlist"
-          />
+          <ViralLoopsButton type="button" text="Join the waitlist" />
           <Button
             type="white-subtle"
             onClick={() => (window.location.href = "/")}
